+++ src/lib.rs
@@ -6,7 +6,7 @@
 //! # cargo_auto_lib
 //!
 //! **Automation tasks coded in Rust language for the workflow of Rust projects**  
-//! ***version: 2.1.4 date: 2024-04-17 author: [Bestia.dev](https://bestia.dev) repository: [GitHub](https://github.com/automation-tasks-rs/cargo_auto_lib)***
+//! ***version: 2.1.7 date: 2024-04-17 author: [Bestia.dev](https://bestia.dev) repository: [GitHub](https://github.com/automation-tasks-rs/cargo_auto_lib)***
 //!
 //!  ![maintained](https://img.shields.io/badge/maintained-green)
 //!  ![ready-for-use](https://img.shields.io/badge/ready_for_use-green)
@@ -26,8 +26,8 @@
 //!  [![Newest docs](https://img.shields.io/badge/newest_docs-blue.svg)](https://automation-tasks-rs.github.io/cargo_auto_lib/cargo_auto_lib/index.html)
 //!  ![cargo_auto_lib](https://bestia.dev/webpage_hit_counter/get_svg_image/276360626.svg)
 //!
-//! [![Lines in Rust code](https://img.shields.io/badge/Lines_in_Rust-1705-green.svg)](https://github.com/automation-tasks-rs/cargo_auto_lib/)
-//! [![Lines in Doc comments](https://img.shields.io/badge/Lines_in_Doc_comments-935-blue.svg)](https://github.com/automation-tasks-rs/cargo_auto_lib/)
+//! [![Lines in Rust code](https://img.shields.io/badge/Lines_in_Rust-1716-green.svg)](https://github.com/automation-tasks-rs/cargo_auto_lib/)
+//! [![Lines in Doc comments](https://img.shields.io/badge/Lines_in_Doc_comments-984-blue.svg)](https://github.com/automation-tasks-rs/cargo_auto_lib/)
 //! [![Lines in Comments](https://img.shields.io/badge/Lines_in_comments-287-purple.svg)](https://github.com/automation-tasks-rs/cargo_auto_lib/)
 //! [![Lines in examples](https://img.shields.io/badge/Lines_in_examples-28-yellow.svg)](https://github.com/automation-tasks-rs/cargo_auto_lib/)
 //! [![Lines in tests](https://img.shields.io/badge/Lines_in_tests-79-orange.svg)](https://github.com/automation-tasks-rs/cargo_auto_lib/)
@@ -139,7 +139,6 @@ mod auto_cargo_toml_mod;
 mod auto_cargo_toml_to_md_mod;
 mod auto_check_micro_xml_mod;
 mod auto_copy_files_to_strings_mod;
-mod auto_crates_io_mod;
 mod auto_delete_old_js_snippets_mod;
 mod auto_doc_tidy_html_mod;
 mod auto_git_mod;
@@ -151,6 +150,7 @@ mod auto_plantuml_mod;
 mod auto_playground_mod;
 mod auto_semver_mod;
 mod auto_semver_or_date_mod;
+mod auto_shell_mod;
 mod auto_version_from_date_mod;
 mod error_mod;
 mod public_api_mod;